# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Monocle",
    "description": "Personal knowledge capture and retrieval API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ingest": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest",
        "description": "Run the 8-step ingest pipeline and return the created note + confidence.",
        "operationId": "ingest_api_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/stream": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest Stream",
        "description": "Streaming ingest \u2014 emits SSE progress events for each pipeline step.",
        "operationId": "ingest_stream_api_ingest_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ingest/failures/retry": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Retry Failure",
        "description": "Retry a previously failed ingest using stored content preview.",
        "operationId": "retry_failure_api_ingest_failures_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Retry Failure Api Ingest Failures Retry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/failures/{failure_id}": {
      "delete": {
        "tags": [
          "ingest"
        ],
        "summary": "Delete Failure",
        "description": "Remove a failed-ingest record from the list.\n\nThe underlying ``.error.md`` sidecar file is NOT deleted.",
        "operationId": "delete_failure_api_ingest_failures__failure_id__delete",
        "parameters": [
          {
            "name": "failure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Failure Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestConfidence": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "template_match": {
            "type": "number",
            "title": "Template Match"
          },
          "metadata_coverage": {
            "type": "number",
            "title": "Metadata Coverage"
          },
          "tag_plausibility": {
            "type": "number",
            "title": "Tag Plausibility"
          },
          "entity_match": {
            "type": "number",
            "title": "Entity Match"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "similar_note_detected": {
            "type": "boolean",
            "title": "Similar Note Detected",
            "default": false
          },
          "similar_note_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similar Note Path"
          }
        },
        "type": "object",
        "required": [
          "score",
          "template_match",
          "metadata_coverage",
          "tag_plausibility",
          "entity_match"
        ],
        "title": "IngestConfidence",
        "description": "Deterministic confidence score for an ingested note (no LLM call)."
      },
      "IngestRequest": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "text/plain"
          },
          "source": {
            "type": "string",
            "enum": [
              "web",
              "voice",
              "teams",
              "mcp",
              "import"
            ],
            "title": "Source",
            "default": "web"
          },
          "template_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Hint"
          },
          "audio_bytes": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Bytes"
          },
          "audio_mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Mime Type"
          },
          "allow_duplicate": {
            "type": "boolean",
            "title": "Allow Duplicate",
            "default": false
          }
        },
        "type": "object",
        "title": "IngestRequest"
      },
      "IngestResponse": {
        "properties": {
          "note": {
            "$ref": "#/components/schemas/Note"
          },
          "confidence": {
            "$ref": "#/components/schemas/IngestConfidence"
          }
        },
        "type": "object",
        "required": [
          "note",
          "confidence"
        ],
        "title": "IngestResponse",
        "description": "Response body for POST /api/ingest."
      },
...
      "LinkRef": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "LinkRef",
        "description": "A structured link from a note's `links:` frontmatter field."
      },
      "Note": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "metadata": {
            "$ref": "#/components/schemas/NoteMetadata"
          },
          "mtime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mtime"
          }
        },
        "type": "object",
        "required": [
          "file_path",
          "title",
          "body"
        ],
        "title": "Note",
        "description": "Full note contents including frontmatter metadata and Markdown body."
      },
      "NoteMetadata": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "person_note",
              "decision",
              "idea",
              "observation",
              "reference",
              "meeting_note",
              "project",
              "action_item",
              "weekly_summary",
              "other"
            ],
            "title": "Type",
            "default": "other"
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": "blank"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "personal"
          },
          "people": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "People"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "source": {
            "type": "string",
            "enum": [
              "web",
              "voice",
              "teams",
              "mcp",
              "import"
            ],
            "title": "Source",
            "default": "web"
          },
          "action_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Items"
          },
          "created": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          },
          "confidence_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Rationale"
          },
          "review_status": {
            "type": "string",
            "enum": [
              "pending",
              "approved"
            ],
            "title": "Review Status",
            "default": "approved"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approval_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "auto",
                  "manual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Mode"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkRef"
            },
            "type": "array",
            "title": "Links"
          },
          "org": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "NoteMetadata",
        "description": "Parsed frontmatter of a note as written/read by the ingest pipeline."
      },
...
      "RetryRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "RetryRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}